==> 04_Recommendations.Rmd <==
# Recommendations {#Recommendations}

Further study is recommended to determine key foraging areas and obtain larger sample sizes to analyze nest site attendance. Increasing the density of shoreline stations throughout the region could highlight currently unknown key foraging areas along the coast, or alternatively, could indicate the Least Terns are foraging either within close proximity to the colony, or farther offshore than stations can detect. Setting up a small omni directional antenna within the colony could provide more robust detections of birds while on the nest. This, coupled with some level of manual tracking to confirm the assumptions made in this report regarding nest site attendance, could further our knowledge of the spatial and temporal requirements of breeding Least Terns.
